Fixed Glitchy Mobs
Made it so that the mobs don’t glitch out when they follow you. The
problem was comparing doubles. They should have been ints.

+++ Desacore/src/pstoriz/desacore/entity/mob/Bandit.java
@@ -42,10 +42,10 @@ public class Bandit extends Mob {
 			xa = 0;
 			ya = 0;
 			Player player = players.get(0);
-			if (x < player.getX()) xa += speed;
-			if (x > player.getX()) xa -= speed;
-			if (y < player.getY()) ya += speed;
-			if (y > player.getY()) ya -= speed;
+			if ((int) x < (int) player.getX()) xa += speed;
+			if ((int) x > (int) player.getX()) xa -= speed;
+			if ((int) y < (int) player.getY()) ya += speed;
+			if ((int) y > (int) player.getY()) ya -= speed;
 		} else {
 			timer++;
 			// time % 60 is once per second
